+++ latte-tests.js
@@ -16,7 +16,7 @@
 'use strict';
 const fs = require('fs');
 const { NAME_CLASH_RE, resolveSuiteFile, extractAnchoredRegex } = require('./tools/repo-checks');
-const EXPECTED_ASSERTIONS = 3163;
+const EXPECTED_ASSERTIONS = 3191;
 
 let file;
 try {